+++ src/zakuro.cpp
@@ -286,6 +286,8 @@ int Zakuro::Run() {
 
     SoraClientConfig sorac_config;
     sorac_config.insecure = config_.insecure;
+    sorac_config.client_cert = config_.client_cert;
+    sorac_config.client_key = config_.client_key;
     sorac_config.signaling_urls = config_.sora_signaling_urls;
     sorac_config.channel_id = config_.sora_channel_id;
     sorac_config.video = config_.sora_video;